backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1
pydantic[email]
passlib[bcrypt]
#Testing Packages
pytest==7.4.0
httpx==0.24.1
#Authentication and Authorization
python-jose==3.3.0
# For authentication
python-multipart==0.0.6
# For service HTML Response
Jinja2==3.1.2
# Service static files
aiofiles==23.1.0
